done csv and finish customer

+++ laravel-app/app/Http/Controllers/CustomerController.php
@@ -2,14 +2,18 @@
 
 namespace App\Http\Controllers;
 
+use App\Exports\CustomerExport;
 use App\Http\Requests\StoreCustomerRequest;
 use App\Http\Requests\UpdateCustomerRequest;
-use App\Imports\CustomerImport;
+use App\Imports\CustomerImportImpl;
+use App\Imports\CustomerImportValidation;
 use App\Models\Customer;
+use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Response;
+use JetBrains\PhpStorm\ArrayShape;
 use Maatwebsite\Excel\Facades\Excel;
-use Maatwebsite\Excel\Validators\ValidationException;
+use mysql_xdevapi\Exception;
 
 class CustomerController extends Controller
 {
@@ -19,15 +23,36 @@ class CustomerController extends Controller
         $this->middleware('auth:api');
     }
 
-    public function index()
+    public function index(Request $request)
     {
+        $page = $request->query('page') + 1 ?? 1;
+        $limit = $request->query('limit') ?? 10;
+        $keyword = $request->query('keyword', '');
+        $sortOrder = $request->query('sort_order', 'desc'); // Default sort order is descending
+
+        $query = Customer::where('user_id', auth()->id())
+            ->when($keyword, function ($q) use ($keyword) {
+                $q->where(function ($innerQ) use ($keyword) {
+                    $innerQ->where('name', 'LIKE', "%$keyword%")
+                        ->orWhere('company', 'LIKE', "%$keyword%")
+                        ->orWhere('email', 'LIKE', "%$keyword%")
+                        ->orWhere('country', 'LIKE', "%$keyword%")
+                        ->orWhere('address', 'LIKE', "%$keyword%")
+                        ->orWhere('contact_number', 'LIKE', "%$keyword%")
+                        ->orWhere('contact_number_country', 'LIKE', "%$keyword%");
+                });
+            })
+            ->orderBy($request->query('sort_by', 'created_at'), $sortOrder); // Sort by the specified column and order
+
+        $customers = $query->simplePaginate($limit, ['*'], 'page', $page);
+
         try {
-            $customers = Customer::all();
-            return Response::customJson(200, $customers, trans('customer.list_success'));
+            return Response::customJson(200, $customers, "success");
         } catch (\Exception $e) {
             return Response::customJson(500, null, $e->getMessage());
         }
     }
+
 
     /**
      * Show the form for creating a new resource.
@@ -45,7 +70,9 @@ class CustomerController extends Controller
     public function store(StoreCustomerRequest $request)
     {
         try {
-            $customer = Customer::create($request->validated());
+            $validatedData = $request->validated();
+            $validatedData['user_id'] = auth()->id();
+            $customer = Customer::create($validatedData);
             return Response::customJson(200, $customer, trans('customer.create_success'));
         } catch (\Exception $e) {
             return Response::customJson(500, null, $e->getMessage());
@@ -82,6 +109,12 @@ class CustomerController extends Controller
         try {
             // Find the customer by ID
             $customer = Customer::findOrFail($id);
+
+            // Check if the customer belongs to the authenticated user
+            if ($customer->user_id !== auth()->user()->id) {
+                return Response::customJson(403, null, "Unauthorized");
+            }
+
             // Update the customer
             $customer->update($request->validated());
 
@@ -90,6 +123,7 @@ class CustomerController extends Controller
             return Response::customJson(500, null, $e->getMessage());
         }
     }
+
 
     /**
      * Remove the specified resource from storage.
@@ -103,7 +137,9 @@ class CustomerController extends Controller
             // Find the customer by ID
             $customer = Customer::findOrFail($id);
 
-            // Delete the customer
+            if ($customer->user_id !== auth()->user()->id) {
+                return Response::customJson(403, null, "Unauthorized");
+            }
             $customer->delete();
 
             return Response::customJson(200, null, trans('customer.delete_success'));
@@ -121,28 +157,65 @@ class CustomerController extends Controller
         try {
             $csvPath = $request->file('csv_file')->getRealPath();
 
-            $reader = Excel::toArray(new CustomerImport(), $csvPath);
-
-            $rows = $reader[0];
-
-            if ($this->isValidCSVFormat($rows[0])) {
-                return Response::customJson(200, null, "CSV is valid");
+            $result = $this->isValidCSVFormat($csvPath);
+            if ($result["valid"]) {
+                return Response::customJson(200, null, "CSV's format is valid");
             }
-            return Response::customJson(400, null, "CSV's format is in valid");
+            return Response::customJson(400, $result, "CSV is not valid");
 
-        } catch (ValidationException $e) {
+        } catch (Exception $e) {
+
             return Response::customJson(500, null, $e->getMessage());
         }
     }
 
-    private function isValidCSVFormat($headerRow)
+    #[ArrayShape(["valid" => "bool", "errors" => "\Illuminate\Support\Collection|\Maatwebsite\Excel\Validators\Failure[]"])] private function isValidCSVFormat($csvPath): array
     {
+        $valid = true;
         $expectedHeaders = ['name', 'company', 'email', 'country', 'address', 'contact_number', 'contact_number_country'];
+        $reader = Excel::toArray(new CustomerImportValidation(), $csvPath);
 
-        if (count($headerRow) !== count($expectedHeaders)) {
-            return false; // Number of columns in header does not match expected headers.
+        $rows = $reader[0];
+        $headerRow = array_keys($rows[0]);
+        $import = new CustomerImportValidation();
+        $import->import($csvPath);
+        $failures = $import->failures();
+
+        if (count($headerRow) != count($expectedHeaders) || $headerRow != $expectedHeaders || count($failures) > 0) {
+            $valid = false;
         }
-
-        return $expectedHeaders == $headerRow;
+        return ["valid" => $valid, "errors" => $failures];
     }
+
+    public function saveCSV(Request $request)
+    {
+        $request->validate([
+            'csv_file' => 'required|file|mimes:csv|max:2048',
+        ]);
+
+        try {
+            $csvPath = $request->file('csv_file')->getRealPath();
+
+
+            $import = new CustomerImportImpl();
+            $import->import($csvPath);
+            if ($import->failures()->isNotEmpty()) {
+                return Response::customJson(400, null, $import->failures());
+            }
+            return Response::customJson(200, null, "Successfully imported");
+        } catch (\Exception $e) {
+            return Response::customJson(500, null, $e->getMessage());
+        }
+    }
+
+    public function exportCsv()
+    {
+        try {
+            $now = Carbon::now()->format('Y-m-d H:i:s');
+            return Excel::download(new CustomerExport(), 'customers_' . $now . '.csv');
+        } catch (\Exception $e) {
+            return Response::customJson(500, null, $e->getMessage());
+        }
+    }
+
 }
